+++ k8s/client/clientset/versioned/scheme/register.go
@@ -6,13 +6,12 @@
 package scheme
 
 import (
+	backupv1 "github.com/bborbe/backup/k8s/apis/backup.benjamin-borbe.de/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	serializer "k8s.io/apimachinery/pkg/runtime/serializer"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-
-	backupv1 "github.com/bborbe/backup/k8s/apis/backup.benjamin-borbe.de/v1"
 )
 
 var Scheme = runtime.NewScheme()