+++ k8s/client/informers/externalversions/backup.benjamin-borbe.de/v1/target.go
@@ -9,15 +9,14 @@ import (
 	context "context"
 	time "time"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	runtime "k8s.io/apimachinery/pkg/runtime"
-	watch "k8s.io/apimachinery/pkg/watch"
-	cache "k8s.io/client-go/tools/cache"
-
 	apisbackupbenjaminborbedev1 "github.com/bborbe/backup/k8s/apis/backup.benjamin-borbe.de/v1"
 	versioned "github.com/bborbe/backup/k8s/client/clientset/versioned"
 	internalinterfaces "github.com/bborbe/backup/k8s/client/informers/externalversions/internalinterfaces"
 	backupbenjaminborbedev1 "github.com/bborbe/backup/k8s/client/listers/backup.benjamin-borbe.de/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	runtime "k8s.io/apimachinery/pkg/runtime"
+	watch "k8s.io/apimachinery/pkg/watch"
+	cache "k8s.io/client-go/tools/cache"
 )
 
 // TargetInformer provides access to a shared informer and lister for
@@ -45,20 +44,32 @@ func NewTargetInformer(client versioned.Interface, namespace string, resyncPerio
 // one. This reduces memory footprint and number of connections to the server.
 func NewFilteredTargetInformer(client versioned.Interface, namespace string, resyncPeriod time.Duration, indexers cache.Indexers, tweakListOptions internalinterfaces.TweakListOptionsFunc) cache.SharedIndexInformer {
 	return cache.NewSharedIndexInformer(
-		&cache.ListWatch{
+		cache.ToListWatcherWithWatchListSemantics(&cache.ListWatch{
 			ListFunc: func(options metav1.ListOptions) (runtime.Object, error) {
 				if tweakListOptions != nil {
 					tweakListOptions(&options)
 				}
-				return client.BackupV1().Targets(namespace).List(context.TODO(), options)
+				return client.BackupV1().Targets(namespace).List(context.Background(), options)
 			},
 			WatchFunc: func(options metav1.ListOptions) (watch.Interface, error) {
 				if tweakListOptions != nil {
 					tweakListOptions(&options)
 				}
-				return client.BackupV1().Targets(namespace).Watch(context.TODO(), options)
+				return client.BackupV1().Targets(namespace).Watch(context.Background(), options)
 			},
-		},
+			ListWithContextFunc: func(ctx context.Context, options metav1.ListOptions) (runtime.Object, error) {
+				if tweakListOptions != nil {
+					tweakListOptions(&options)
+				}
+				return client.BackupV1().Targets(namespace).List(ctx, options)
+			},
+			WatchFuncWithContext: func(ctx context.Context, options metav1.ListOptions) (watch.Interface, error) {
+				if tweakListOptions != nil {
+					tweakListOptions(&options)
+				}
+				return client.BackupV1().Targets(namespace).Watch(ctx, options)
+			},
+		}, client),
 		&apisbackupbenjaminborbedev1.Target{},
 		resyncPeriod,
 		indexers,